--- persistence/nosql/nodes/spi/src/main/java/org/apache/polaris/nosql/nodeids/spi/NodeStore.java
@@ -1,34 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.polaris.nosql.nodeids.spi;
-
-import jakarta.annotation.Nonnull;
-import jakarta.annotation.Nullable;
-import java.util.Optional;
-
-/** Abstraction of nodes persistence for the node management implementation. */
-public interface NodeStore {
-  Optional<NodeState> fetch(int nodeId);
-
-  @Nonnull
-  NodeState[] fetchMany(@Nonnull int... nodeIds);
-
-  @Nullable
-  NodeState persist(int nodeId, Optional<NodeState> expectedNodeState, @Nonnull NodeState newState);
-}
